fix : 프로젝트 전체 구조 수정
