+++ src/App.tsx
@@ -2,19 +2,18 @@ import React, { useEffect, useState } from 'react';
 import './index.css';
 import { Redirect, Route, Switch } from 'react-router';
 import Home from '@pages/home';
-import Result from '@pages/result';
 import Layout from '@components/layouts';
-import SearchNewsService from '@services/search/search-news';
 import AuthService from '@services/auth/auth-service';
 import { useSetRecoilState } from 'recoil';
 import { userState } from '@stores/user';
+import PostWrite from '@pages/post/write';
+import PostDetail from '@pages/post/detail';
 
 type Props = {
-	searchNewsService: SearchNewsService;
 	authService: AuthService;
 };
 
-const App = ({ searchNewsService, authService }: Props) => {
+const App = ({ authService }: Props) => {
 	const [isUserChecked, setIsUserChecked] = useState(false);
 	const setUser = useSetRecoilState(userState);
 
@@ -30,7 +29,6 @@ const App = ({ searchNewsService, authService }: Props) => {
 						...prev,
 						isLogin: true,
 						id: result.data!.userId,
-						recentKeywords: result.data!.Keywords.slice(0, 5),
 					};
 				});
 			} else {
@@ -39,7 +37,6 @@ const App = ({ searchNewsService, authService }: Props) => {
 						...prev,
 						isLogin: false,
 						id: '',
-						recentKeywords: [],
 					};
 				});
 			}
@@ -60,10 +57,13 @@ const App = ({ searchNewsService, authService }: Props) => {
 				<Layout authService={authService}>
 					<Switch>
 						<Route exact path="/">
-							<Home searchNewsService={searchNewsService} />
+							<Home />
 						</Route>
-						<Route exact path="/result/:keyword">
-							<Result searchNewsService={searchNewsService} />
+						<Route exact path="post/write">
+							<PostWrite />
+						</Route>
+						<Route exact path="post/:id">
+							<PostDetail />
 						</Route>
 						<Redirect path="*" to="/" />
 					</Switch>
