--- src/components/features/home/recent-keywords/tests/RecentKeywords.test.tsx
@@ -1,50 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import renderer from 'react-test-renderer';
-import RecentKeywords from '@components/features/home/recent-keywords';
-import userEvent from '@testing-library/user-event';
-import { Keyword } from 'src/types/home/types';
-
-describe('RecentKeywords', () => {
-	const moveToReultPage = jest.fn();
-
-	const mockRecentKeywords: Keyword[] = [
-		{
-			id: 1,
-			content: 'korea',
-		},
-		{
-			id: 2,
-			content: 'seoul',
-		},
-	];
-
-	beforeEach(() => {
-		render(<RecentKeywords recentKeywords={mockRecentKeywords} moveToReultPage={moveToReultPage} />);
-	});
-
-	it('Snapshot', () => {
-		const component = renderer.create(
-			<RecentKeywords recentKeywords={mockRecentKeywords} moveToReultPage={moveToReultPage} />,
-		);
-		expect(component.toJSON()).toMatchSnapshot();
-	});
-
-	it('Show recent keyword list delivered through props', () => {
-		const keywords = screen.getAllByRole('listitem');
-
-		expect(keywords).toHaveLength(2);
-
-		expect(keywords[0]).toHaveTextContent('korea');
-
-		expect(keywords[1]).toHaveTextContent('seoul');
-	});
-
-	it('Move to result page when click keyword', () => {
-		const keywords = screen.getAllByRole('listitem');
-
-		userEvent.click(keywords[0]);
-
-		expect(moveToReultPage).toHaveBeenCalledWith('korea');
-	});
-});
