+++ src/pages/post/detail/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const PostDetail = () => {
+	return (
+		<div>
+			<>detail</>
+		</div>
+	);
+};
+
+export default PostDetail;
