+++ src/pages/post/write/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const PostWrite = () => {
+	return (
+		<div>
+			<>write</>
+		</div>
+	);
+};
+
+export default PostWrite;
