+++ src/services/post/post-service.ts
@@ -1,0 +1,8 @@
+import axios, { AxiosInstance } from 'axios';
+import API_URL, { API_HOST, BasicResult } from '@services/api';
+
+class SearchNewsService {
+	constructor() {}
+}
+
+export default SearchNewsService;
