+++ src/services/post/types.ts
@@ -1,0 +1,6 @@
+export type Post = {
+	id: number;
+	title: string;
+	content: string;
+	createdAt: string;
+};
