--- src/services/search/types.ts
@@ -1,3 +1,0 @@
-export type AddRecentKeywordRequest = {
-	content: string;
-};
