--- src/types/home/types.ts
@@ -1,4 +1,0 @@
-export type Keyword = {
-	id: number;
-	content: string;
-};
